% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\docType{data}
\name{anthro1}
\alias{anthro1}
\title{Anthropometric data from a SMART survey in Kabul, Afghanistan.}
\format{A data frame with 873 observations and 11 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{age}}{Age of child (months)}
\item{\code{sex}}{Gender of child}
\item{\code{weight}}{Weight of child (kgs)}
\item{\code{height}}{Height of child (cm)}
\item{\code{muac}}{Mid-upper arm circumference (mm)}
\item{\code{oedema}}{Presence or absence of oedema}
\item{\code{haz}}{Height-for-age z-score}
\item{\code{waz}}{Weight-for-age z-score}
\item{\code{whz}}{Weight-for-height z-score}
\item{\code{flag}}{Data quality flag}
}}
\usage{
anthro1
}
\description{
Anthropometric data from a SMART survey in Kabul, Afghanistan.
}
\keyword{datasets}
