% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\docType{data}
\name{anthro2}
\alias{anthro2}
\title{Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.}
\format{A data frame with 796 observations and 6 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender}
\item{\code{wt}}{Weight (kg)}
\item{\code{ht}}{height (cm)}
\item{\code{oedema}}{Presence or absence of oedema}
}}
\usage{
anthro2
}
\description{
Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.
}
\keyword{datasets}
