% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro3}
\alias{anthro3}
\title{Anthropometric data from a Rapid Assessment Method (RAM) survey from Burundi.}
\format{
A data frame with 221 observations and 7 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Definition} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender \cr
   \emph{weight} \tab Weight (kg) \cr
   \emph{height} \tab Height (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (cm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
anthro3
}
\description{
Anthropometric data from a Rapid Assessment Method (RAM) survey from Burundi.
}
\keyword{datasets}
