% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro4}
\alias{anthro4}
\title{A subset of mid-upper arm circumference data from study conducted to create
MUAC-for-age z-scores}
\format{
A data.frame with 257 observations and 4 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Definition} \cr
   \emph{pk_serial} \tab Unique identifier \cr
   \emph{muac} \tab Mid-upper arm circumference in centimetres \cr
   \emph{agemons} \tab Age in months \cr
   \emph{sex} \tab Sex; 1 = Male; 2 = Female \cr
}
}
\source{
Mramba Lazarus, Ngari Moses, Mwangome Martha, Muchai Lilian, Bauni
Evasius, Walker A Sarah et al. A growth reference for mid upper arm
circumference for age among school age children and adolescents, and
validation for mortality: growth curve construction and longitudinal
cohort study BMJ 2017; 358 :j3423 https://doi.org/10.1136/bmj.j3423
}
\usage{
anthro4
}
\description{
A subset of mid-upper arm circumference data from study conducted to create
MUAC-for-age z-scores
}
\keyword{datasets}
