% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi}
\alias{calculate_bmi}
\title{calculate_bmi}
\usage{
calculate_bmi(weight, height)
}
\arguments{
\item{weight}{A numeric value for weight in kilograms.}

\item{height}{A numeric value for height in metres.}
}
\value{
A numeric value for body mass index (BMI) in \code{kgs / m ^ 2}
}
\description{
Function to calculate body mass index (BMI) using weight (in kgs) and
height (in metres)
}
\examples{
# Calculate BMI of a child with weight 14.6 kgs and a height of 98.0 cms
calculate_bmi(weight = 14.6, height = 98.0 / 100)

}
