% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_bmi.R
\name{calculate_bmi_cohort}
\alias{calculate_bmi_cohort}
\title{calculate_bmi_cohort}
\usage{
calculate_bmi_cohort(data = NULL, weight, height)
}
\arguments{
\item{data}{Data frame containing corresponding data on \code{weight} and
\code{height} of children. Default is NULL. If specified, parameters for
\code{weight} and \code{height} should be provided as character values of the names of
variables in \code{data} corresponding to the parameters required.}

\item{weight}{Either numeric values for weight in kg with at least 1 decimal
place (default) or character value (if \code{data} is specified) indicating
variable name in \code{data} containing information on weight of
child/children.}

\item{height}{Either numeric values for height in cm with at least 1 decimal
place (default) or character value (if \code{data} is specified) indicating
variable name in \code{data} containing information on height of
child/children.}
}
\value{
Either a single numeric value for BMI if data is for single child or a
    data frame of numeric values for BMI if data is for multiple children.
}
\description{
Function to calculate body mass index for a dataset containing weight (in kgs)
and height (in metres). A wrapper function using \code{mapply()} to vectorise
application of \code{calculate_bmi()} to calculate BMI for each row in dataset.
}
\examples{
# apply \\code{calculate_bmi_cohort} to \\code{anthro1} dataset
calculate_bmi_cohort(weight = anthro1$weight,
                     height = (anthro1$height / 100) ^ 2)

}
