% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addCDC.R
\name{getCDC}
\alias{getCDC}
\alias{addCDC}
\title{Calculate CDC Growth Reference z-score for given anthropometric
measurement/s.}
\usage{
getCDC(sex, firstPart, secondPart, thirdPart = NA, index = NA, standing = NA)

addCDC(
  data,
  sex,
  firstPart,
  secondPart,
  thirdPart = NA,
  index = NA,
  standing = NULL,
  output = paste(index, "z", sep = ""),
  digits = 2
)
}
\arguments{
\item{sex}{A vector specifying the sex of the subject or a character value
for the name of the variable in \code{data} specifying the sex of the subject.
This must be coded as \code{1 = male} and \code{2 = female}. Give a quoted variable
name as in (e.g.) \code{"sex"}.}

\item{firstPart}{A vector or a character value for the name of the variable
in \code{data} specifying:
\itemize{
\item Weight (kg) for BMI/A, W/A, W/H, or W/L
\item Head circumference (cm) for HC/A
\item Height (cm) for BMI/A for H/A
\item Length (cm) for L/A
}

If name of variable in \code{data}, give a quoted variable name as in (e.g.)
\code{"weight"}. Be careful with units (weight in kg; height, length, head
circumference in cm).}

\item{secondPart}{A vector or a character value for the name of variable
in \code{data} specifying:
\itemize{
\item Age (days) for H/A, HC/A, or L/A
\item Height (cm) BMI/A or W/H
\item Length (cm) for W/L
}

If name of variable in \code{data}, give a quoted variable name as in (e.g.)
\code{"age"}. Be careful with units (age in months; height and length in cm).}

\item{thirdPart}{A vector or a character value for the name of variable
in \code{data} specifying age (in months) for BMI/A. If name of variable in
\code{data}, give a quoted variable name as in (e.g.) \code{"age"}. Be careful with
units (age in months).}

\item{index}{The index to be calculated. One of:\tabular{ll}{
   \strong{Index} \tab \strong{Definition} \cr
   \code{bfa} \tab BMI for age \cr
   \code{hca} \tab Head circumference for age \cr
   \code{hfa} \tab Height for age \cr
   \code{lfa} \tab Length for age \cr
   \code{wfa} \tab Weight for age \cr
   \code{wfh} \tab Weight for height \cr
   \code{wfl} \tab Weight for length \cr
}


Give a quoted index name as in (e.g.) \code{"wfh"}.}

\item{standing}{A vector or a character value for name of variable in \code{data}
specifying how stature was measured. If NULL then age (for \code{"hfa"} or
\code{"lfa"}) or height rules (for \code{"wfh"} or \code{"wfl"}) will be applied.
This must be coded as \code{1 = Standing}; \code{2 = Supine}; \code{3 = Unknown}. All
other values will be recoded to \code{3 = Unknown}. If name of variable in
\code{data}, give a quoted variable name as in (e.g.) \code{"measured"} or a single
value (e.g. \verb{"measured = 1}). If no value (or NULL) is specified then age
rules will be applied.}

\item{data}{A survey dataset as a data.frame object}

\item{output}{The name of the column containing the specified index to be
added to the dataset. This is an optional parameter. If you do not specify
a value for output then the added column will take the name of the
specified index with a \code{"z"} appended.}

\item{digits}{The number of decimal places for \code{output}. Defaults to 2 d.p.}
}
\value{
A data.frame of the survey dataset with the calculated z-scores
added.
}
\description{
The first function, \code{\link[=getCDC]{getCDC()}}, is usually called by the \code{\link[=addCDC]{addCDC()}}
function but could be used as a stand-alone calculator for getting z-score
for a given anthropometric measurement.
}
\details{
\code{\link[=addCDC]{addCDC()}} adds the CDC Growth Reference z-scores to a data frame of
anthropometric data for weight, height or length, head circumference, and
body mass index (BMI)
}
\examples{
# Given a male child 10 months old with a weight of 5.7 kgs, height of 64.2
# cms, and MUAC of 125 mm:
#
# Calculate weight-for-height
getCDC(sex = 1,
       firstPart = 5.7,
       secondPart = 64.2,
       index = "wfh",
       standing = 3)

# calculate weight-for-age
getCDC(sex = 1,
       firstPart = 5.7,
       secondPart = 10,
       index = "wfa",
       standing = 3)

# calculate height-for-age
getCDC(sex = 1,
       firstPart = 64.2,
       secondPart = 10,
       index = "hfa",
       standing = 3)

# Calculate weight-for-height (wfh) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "weight",
       secondPart = "height",
       index = "wfh")

# Calculate weight-for-age (wfa) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "weight",
       secondPart = "age",
       index = "wfa")

# Calculate height-for-age (hfa) for the anthro3 dataset
addCDC(data = anthro3,
       sex = "sex",
       firstPart = "height",
       secondPart = "age",
       index = "hfa")

}
