% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cdcData}
\alias{cdcData}
\title{United States Center for Disease Control (CDC) Growth Charts Data}
\format{
A data frame with 6 columns and 1673 rows.\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{indicator} \tab One of weight-for-age (\code{waz}), height-for-age (\code{haz}), or \cr
    \tab weight-for-height (\code{whz}) anthropometric indicators \cr
   \emph{sex} \tab Sex of child (1 = Male; 2 = Female) \cr
   \emph{given} \tab Variable to which standardisation is to be made. For \cr
    \tab \code{waz} and \code{haz}, \code{given} is age in months. For \code{whz}, \code{given} is height \cr
    \tab in cm \cr
   \emph{l} \tab \code{L} component of the LMS method for normalising growth \cr
    \tab centile standards. \code{L} is the trend in the optimal power to obtain \cr
    \tab normality \cr
   \emph{m} \tab \code{M} component of the LMS method for normalising growth centile \cr
    \tab standards. \code{M} is the trend in the mean \cr
   \emph{s} \tab \code{S} component of the LMS method for normalising growth centile \cr
    \tab standards. \code{S} is the trend in the coefficient of variation \cr
}
}
\source{
https://www.cdc.gov/growthcharts/percentile_data_files.htm
}
\usage{
cdcData
}
\description{
United States Center for Disease Control (CDC) Growth Charts Data
}
\keyword{datasets}
