% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_undernutrition.R
\name{find_child_underweight}
\alias{find_child_underweight}
\title{Find children who are underweight from an anthropometric dataset}
\usage{
find_child_underweight(df, zscore = "waz", flag = NULL, add = TRUE)
}
\arguments{
\item{df}{A data.frame containing anthropometric z-score indices for
`weight-for-age`}

\item{zscore}{A character value for variable name for `weight-for-age z-score`
in `df`. Default is `waz`}

\item{flag}{A character value for variable name of coded flags applied to
`weight-for-age z-score` values in `df`. Default is NULL indicating
that flags have already been applied and flagged values have been censored.}

\item{add}{Logical. Should recode values for underweight be added to
`df`? Default is TRUE}
}
\value{
if `add` TRUE, returns `df` with recode values indicating underweight
  cases.
}
\description{
Find children who are underweight from an anthropometric dataset
}
\examples{

find_child_underweight(df = anthro1, zscore = "waz", flag = NULL, add = TRUE)

}
