% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acfa_chart_data.R
\name{get_acfa_zchart}
\alias{get_acfa_zchart}
\title{get_acfa_zchart}
\usage{
get_acfa_zchart(baseurl = "http://www.who.int/childgrowth/standards/",
  gender = c("boys", "girls"))
}
\arguments{
\item{baseurl}{Character value for the base URL of the WHO Child Growth
Standards. The default is \url{http://www.who.int/childgrowth/standards/}.
This can be adjusted should WHO change the base URL for the Child Growth
Standards section of their website.}

\item{gender}{A character value or vector that specifies the sex-specific
reference data to extract. The default is \code{c("boys", "girls")}.}
}
\value{
A data frame in tidy format with 5 coloumns and 31788 rows:
\describe{
\item{\code{sex}}{Sex of child. 1 = boy; 2 = girl}
\item{\code{month}}{Age of child in months. This is calculated from the
    original data using that data for age of child in days}
\item{\code{day}}{Age of child in days. This is provided for by default in
    the original data}
\item{\code{sd_type}}{Type of z-score. Can be one of \code{4SD}, \code{3SD},
    \code{2SD}, \code{1SD}, \code{0}, \code{-1SD}, \code{-2SD}, \code{-3SD},
    \code{-4SD}}
\item{\code{sd_value}}{Arm circumference (in centimetres) value for the specified
    \code{sd_type}}
}
}
\description{
Function to extract arm circumference-for-age z-score data from the World Health
Organization (WHO) website on Child Growth Standards
\url{http://www.who.int/childgrowth/standards/"} which contains a
collection of reference data from the WHO Multicentre Growth Reference Study
(MGRS). This function extracts the expanded data provided by WHO for creating
growth charts for arm circumference-for-age z-scores.
}
\examples{
# Get arm circumference-for-age z-score expanded tables for charting
get_acfa_zchart()

}
