% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_centiles_all.R
\name{get_centiles_all}
\alias{get_centiles_all}
\title{get_centiles_all}
\usage{
get_centiles_all(data = NULL, sex, weight = NULL, height = NULL,
  bmi = NULL, hc = NULL, ac = NULL, ss = NULL, ts = NULL, age, index)
}
\arguments{
\item{data}{Data frame containing corresponding data on \code{sex}, \code{age},
\code{weight}, \code{height}, \code{BMI}, \code{head circumference},
\code{arm circumference}, \code{subscapular skinfold} and
\code{triceps skinfold} of children. Default is NULL. If specified,
parameters for \code{sex}, \code{age}, \code{weight}, \code{height},
\code{BMI}, \code{head circumference}, \code{arm circumference},
\code{subscapular skinfold} and \code{triceps skinfold} should be provided
as character values of the names of variables in \code{data} corresponding
to the parameters required.}

\item{sex}{Either numeric values (1 = male; 2 = female) indicating sex of
child (default) or character value (if \code{data} is specified) indicating
variable name in \code{data} containing information on sex of
child/children (1 = male; 2 = female).}

\item{weight}{Either numeric values for weight in kg with at least 1 decimal
place (default) or character value (if \code{data} is specified) indicating
variable name in \code{data} containing information on weight of
child/children.}

\item{height}{Either numeric values for height in cm with at least 1 decimal
place (default) or character value (if \code{data} is specified) indicating
variable name in \code{data} containing information on height of
child/children.}

\item{bmi}{Either numeric values for BMI in \eqn{kg / m ^ 2} (default) or
character value (if \code{data} is specified) indicating variable name in
\code{data} containing information on BMI of child/children.}

\item{hc}{Either numeric values for head circumference in cm (default) or
character value (if \code{data} is specified) indicating variable name in
\code{data} containing information on head circumference of child/children.}

\item{ac}{Either numeric values for arm circumference in cm (default) or
character value (if \code{data} is specified) indicating variable name in
\code{data} containing information on arm circumference of child/children.}

\item{ss}{Either numeric values for subscapular skinfold in cm (default) or
character value (if \code{data} is specified) indicating variable name in
\code{data} containing information on subscapular skinfold of child/children.}

\item{ts}{Either numeric values for triceps skinfold in cm (default) or
character value (if \code{data} is specified) indicating variable name in
\code{data} containing information on triceps skinfold of child/children.}

\item{age}{Either numeric values for age in whole months (default) or character
value (if \code{data} is specified) indicating variable name in \code{data}
containing information on age of child/children.}

\item{index}{One of "wfh", "hfa", "wfa", "bfa", "acfa", "hcfa", "ssfa", "tsfa"
(specifies the required index) or "all" to calculate all eight indices}
}
\value{
Either a single numeric value for z-score of the anthropometric index
    selected if data is for single child or a data frame of numeric values for
    z-scores of each anthropometric index if data is for multiple children and
    more than one anthropometric index selected.
}
\description{
Calculate centiles for length/height-for-age (HAZ), weight-for-age (WAZ),
weight-for-height/length (WHZ), BMI-for-age (BAZ), head circumference-for-age
(HCAZ), arm circumference-for-age (ACAZ), subscapular skinfold-for-age (SSAZ)
and triceps skinfold-for-age (TSAZ) using the WHO Growth Reference (2006)
for a cohort of children.
}
\section{Warning}{

The function fails messily when \code{secondPart} is outside of the range
given in the WGS reference (i.e. 45 to 120 cm for height and 0 to 60 months
for age). It is up to you to check the ranges of your data.
}

\examples{
  # apply \\code{get_centiles_all()} to a make believe 52 month old male child with
  # weight of 14.6 kg and height of 98.0 cm
  waz <- get_centiles_all(sex = 1,          # 1 = Male / 2 = Female
                          weight = 14.6,    # Weight in kilograms
                          height = 98,      # Height in centimetres
                          age = 52,         # Age in whole months
                          index = "wfa")    # Anthropometric index (weight-for-age)
  waz

  haz <- get_centiles_all(sex = 1,
                          weight = 14.6,
                          height = 98,      # Height in centimetres
                          age = 52,
                          index = "hfa")    # Anthropometric index (height-for-age)
  haz

  whz <- get_centiles_all(sex = 1,
                          weight = 14.6,
                          height = 98,
                          age = 52,
                          index = "wfh")    # Anthropometric index (weight-for-height)
  whz

  # apply \\code{get_centiles_all()} to \\code{anthro1} dataset
  waz <- get_centiles_all(data = anthro1,
                          sex = "sex",
                          weight = "weight",
                          height = "height",
                          age = "age",
                          index = "wfa")
  waz

  haz <- get_centiles_all(sex = anthro1$sex,
                          weight = anthro1$weight,
                          height = anthro1$height,
                          age = anthro1$age,
                          index = "hfa")
  haz

  all <- get_centiles_all(data = anthro1,
                          sex = "sex",
                          weight = "weight",
                          height = "height",
                          age = "age",
                          index = "all")
  all

}
