% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfa_chart_data.R
\name{get_wfa_pchart}
\alias{get_wfa_pchart}
\title{get_wfa_pchart}
\usage{
get_wfa_pchart(baseurl = "http://www.who.int/childgrowth/standards/",
  gender = c("boys", "girls"))
}
\arguments{
\item{baseurl}{Character value for the base URL of the WHO Child Growth
Standards. The default is \url{http://www.who.int/childgrowth/standards/}.
This can be adjusted should WHO change the base URL for the Child Growth
Standards section of their website.}

\item{gender}{A character value or vector that specifies the sex-specific
reference data to extract. The default is \code{c("boys", "girls")}.}
}
\value{
A data frame in tidy format with 8 coloumns and 55710 rows:
\describe{
\item{\code{sex}}{Sex of child. 1 = boy; 2 = girl}
\item{\code{month}}{Age of child in months. This is calculated from the
    original data using that data for age of child in days}
\item{\code{day}}{Age of child in days. This is provided for by default in
    the original data}
\item{\code{l}}{\code{L} component of the LMS method for normalising growth
    centile standards. \code{L} is the trend in the optimal power to obtain
    normality}
\item{\code{m}}{\code{M} component of the LMS method for normalising growth
    centile standards. \code{M} is the trend in the mean}
\item{\code{s}}{\code{S} component of the LMS method for normalising growth
    centile standards. \code{S} is the trend in the coefficient of variation}
\item{\code{p_type}}{Type of z-score. Can be one of \code{0.10th}, \code{1st},
    \code{3rd}, \code{5th}, \code{10th}, \code{15th}, \code{25th}, \code{50th},
    \code{75th}, \code{85th}, \code{90th}, \code{95th}, \code{97th},
    \code{99th} and \code{99.9th}}
\item{\code{p_value}}{Weight (kgs) value for the specified \code{p_type}}
}
}
\description{
Function to extract weight-for-age percentile data from the World Health
Organization (WHO) website on Child Growth Standards
\url{http://www.who.int/childgrowth/standards/"} which contains a
collection of reference data from the WHO Multicentre Growth Reference Study
(MGRS). This function extracts the expanded data provided by WHO for creating
growth charts for weight-for-age percentiles.
}
\examples{
# Get weight-for-age percentile expanded tables for charting
get_wfa_pchart()

}
