% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wfh_chart_data.R
\name{get_wfh_zchart}
\alias{get_wfh_zchart}
\title{get_wfh_zchart}
\usage{
get_wfh_zchart(baseurl = "http://www.who.int/childgrowth/standards/",
  gender = c("boys", "girls"))
}
\arguments{
\item{baseurl}{Character value for the base URL of the WHO Child Growth
Standards. The default is \url{http://www.who.int/childgrowth/standards/}.
This can be adjusted should WHO change the base URL for the Child Growth
Standards section of their website.}

\item{gender}{A character value or vector that specifies the sex-specific
reference data to extract. The default is \code{c("boys", "girls")}.}
}
\value{
A data frame in tidy format with 4 coloumns and 21636 rows:
\describe{
\item{\code{sex}}{Sex of child. 1 = boy; 2 = girl}
\item{\code{lh}}{Length/height of child in centimetres.}
\item{\code{sd_type}}{Type of z-score. Can be one of \code{4SD}, \code{3SD},
    \code{2SD}, \code{1SD}, \code{0}, \code{-1SD}, \code{-2SD}, \code{-3SD},
    \code{-4SD}}
\item{\code{sd_value}}{Weight (kgs) value for the specified \code{sd_type}}
}
}
\description{
Function to extract weight-for-length/height z-score data from the World Health
Organization (WHO) website on Child Growth Standards
\url{http://www.who.int/childgrowth/standards/"} which contains a
collection of reference data from the WHO Multicentre Growth Reference Study
(MGRS). This function extracts the expanded data provided by WHO for creating
growth charts for weight-for-length/height z-scores.
}
\examples{
# Get weight-for-length/height z-score expanded tables for charting
get_wfh_zchart()

}
