% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_z.R
\name{get_z}
\alias{get_z}
\title{get_z}
\usage{
get_z(sexObserved, firstPart, secondPart, index)
}
\arguments{
\item{sexObserved}{Sex of child (1 = Male; 2 = Female)}

\item{firstPart}{Weight (in kg for WHZ and WAZ) or height (in cm for HAZ) or
BMI (in kg/m squared for BAZ) or head circumference (in cm for HCAZ) or
arm circumference (in cm for ACAZ) or subscapular skinfold (in cm for SSAZ)
or triceps skinfold (in cm for TSAZ)}

\item{secondPart}{Age (in months for HAZ, WAZ, BAZ, HCAZ, ACAZ, SSAZ and TSAZ)
or height (in cm for WHZ)}

\item{index}{One of "wfh", "hfa", "wfa", "bfa", "hcfa", "acfa", "ssfa", "tsfa"
(specifies the required index)}
}
\value{
z-score of the anthropometric index selected
}
\description{
Calculate z-scores for length/height-for-age (HAZ), weight-for-age (WAZ),
weight-for-height/length (WHZ), BMI-for-age (BAZ), head circumference-for-age
(HCAZ), arm circumference-for-age (ACAZ), subscapular skinfold-for-age (SSAZ)
and triceps skinfold-for-age (TSAZ) using the WHO Growth Reference (2006)
for a single child data.
}
\examples{
  # apply \\code{get_z()} to a make believe 52 month old male child with weight of
  # 14.6 kg and height of 98.0 cm
  waz <- get_z(sexObserved = 1,     # 1 = Male / 2 = Female
               firstPart = 14.6,    # Weight in kilograms
               secondPart = 52,     # Age in whole months
               index = "wfa")       # Anthropometric index (weight-for-age)
  waz

  haz <- get_z(sexObserved = 1,
               firstPart = 98,      # Height in centimetres
               secondPart = 52,
               index = "hfa")       # Anthropometric index (height-for-age)
  haz

  whz <- get_z(sexObserved = 1,
               firstPart = 14.6,
               secondPart = 98,
               index = "wfh")       # Anthropometric index (weight-for-height)
  whz

}
