% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWGS.R
\name{getWGS}
\alias{getWGS}
\title{getWGS}
\usage{
getWGS(sexObserved, firstPart, secondPart, index)
}
\arguments{
\item{sexObserved}{Sex of child (1 = Male; 2 = Female)}

\item{firstPart}{Weight (kg; for WHZ and WAZ) or height (cm; for HAZ)}

\item{secondPart}{Age (months; for HAZ and WAZ) or height (cm; for WHZ)}

\item{index}{One of "wfh", "hfa", "wfa" (specified the required index)}
}
\value{
Numeric vector of corresponding z-scores of the anthropometric
    index selected
}
\description{
Function to calculate z-scores for WHZ, HAZ, WAZ using the WHO Growth
Reference (2006) using data and methods from ISBN ISBN 92 4 154693 X.
}
\details{
The function fails messily when "secondPart" is outside of the range given
in the WGS reference (i.e. 45 to 120 cm for height and 0 to 60 months for
age). It is up to you to check the ranges of your data.

The reference data for W/H assumes supine length is used for children with a
standing height below 85cm

Heights should be specified in cm to the nearest mm (i.e. to 1 d.p.)

Ages should be specified in whole months

Weights should be specified in kg to available precision

The function requires reference data \code{wgsData} included in this package
}
\examples{
  # apply getWGS to sample data anthro1
  waz <- getWGS(sexObserved = anthro1$sex[1],
                firstPart = anthro1$weight[1],
                secondPart = anthro1$age[1],
                index = "wfa")

  haz <- getWGS(sexObserved = anthro1$sex[1],
                firstPart = anthro1$height[1],
                secondPart = anthro1$age[1],
                index = "hfa")

  whz <- getWGS(sexObserved = anthro1$sex[1],
                firstPart = anthro1$weight[1],
                secondPart = anthro1$height[1],
                index = "wfh")
}
