% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nhanes}
\alias{nhanes}
\title{A subset of data from the National Health and Nutrition Examination Survey}
\format{
A data.frame with 339 observations and 10 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Definition} \cr
   \emph{cid} \tab Unique child identifier \cr
   \emph{sex} \tab Sex of child. 1 = male; 2 = female \cr
   \emph{agemos} \tab Age in months \cr
   \emph{weight} \tab Weight in kilograms \cr
   \emph{height} \tab Height in centimetres \cr
   \emph{headcir} \tab Head circumference in centimetres \cr
   \emph{waz} \tab Weight-for-age z-score \cr
   \emph{bmiz} \tab BMI-for-age z-score \cr
}
}
\usage{
nhanes
}
\description{
A subset of data from the National Health and Nutrition Examination Survey
}
\examples{
head(nhanes)

}
\keyword{datasets}
