% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\name{zscorer-deprecated}
\alias{zscorer-deprecated}
\title{Deprecated functions in zscorer}
\description{
These functions still work but will be removed (defunct) in the next version.
}
\details{
\itemize{
\item \code{\link[=getWGS]{getWGS()}}
\item \code{\link[=getCohortWGS]{getCohortWGS()}}
\item \code{\link[=getAllWGS]{getAllWGS()}}
}
}
