% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\docType{package}
\name{zscorer}
\alias{zscorer}
\alias{zscorer-package}
\title{zscorer}
\description{
Tool for calculating z-scores for WHZ, HAZ, WAZ using the WHO Growth
Reference (2006) using data and methods from:
}
\details{
\cite{World Health Organization. WHO Child Growth Standards: Length/Height-for-age,
Weight-for-age, Weight-for-length, Weight-for-height, and Body Mass Index-for age:
Methods and Development. 1st ed. World Health Organization; 2006.
ISBN ISBN 92 4 154693 X}
}
