% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addWGSR.R
\name{addWGSR}
\alias{addWGSR}
\title{addWGSR}
\usage{
addWGSR(data, sex, firstPart, secondPart, thirdPart = NA, index = NA,
  standing = NULL, output = paste(index, "z", sep = ""), digits = 2)
}
\arguments{
\item{data}{A survey dataset as a data.frame object}

\item{sex}{Name of variable specifying the sex of the subject. This must be
    coded as 1 = male; 2 = female.

    Give a quoted variable name as in (e.g.) "sex".}

\item{firstPart}{Name of variable specifying:

    \itemize{
      \item Weight (kg) for BMI/A, W/A, W/H, or W/L
      \item Head circumference (cm) for HC/A
      \item Height (cm) for BMI/A for H/A
      \item Length (cm) for L/A
      \item MUAC (cm) for MUAC/A
      \item Sub-scapular skinfold (mm) for SSF/A
      \item Triceps skinfold (mm) for TSF/A
    }

    Give a quoted variable name as in (e.g.) "weight".

    Be careful with units (weight in kg; height, length, head circumference,
    and MUAC in cm, skinfolds in mm).}

\item{secondPart}{Name of variable specifying:

    \itemize{
      \item Age (days) for H/A, HC/A, L/A, MUAC/A, SSF/A, or TSF/A
      \item Height (cm) BMI/A or W/H
      \item Lenght (cm) for W/L
    }

    Give a quoted variable name as in (e.g.) "age".

    Be careful with units (age in days; height and length in cm).}

\item{thirdPart}{Name of varible specifying age (in days) for BMI/A

    Give a quoted variable name as in (e.g.) "age".

    Be careful with units (age in days).}

\item{index}{The index to be calculated and added to \code{data}. One of:

    \describe{
      \item{\code{bfa}}{BMI for age}
      \item{\code{hca}}{Head circumgerence for age}
      \item{\code{hfa}}{Height for age}
      \item{\code{lfa}}{Length for age}
      \item{\code{mfa}}{MUAC for age}
      \item{\code{ssa}}{Sub-scapular skinfold for age}
      \item{\code{tsa}}{Triceps skinfold for age}
      \item{\code{wfa}}{Weight for age}
      \item{\code{wfh}}{Weight for height}
      \item{\code{wfl}}{Weight for length}
    }

    Give a quoted index name as in (e.g.) "wfh".}

\item{standing}{Variable specifying how stature was measured. If NULL then age
    (for "hfa" or "lfa") or height rules (for "wfh" or "wfl") will be applied.

    This must be coded as 1 = Standing; 2 = Supine; 3 = Unknown. All other
    values will be recoded to 3 = Unknown.

    Give a quoted variable name as in (e.g.) "measured" or a single value
    (e.g. "measured = 1"). If no value (or NULL) is specified then height and
    age rules will be applied.}

\item{output}{The name of the column containing the specified index to be
added to the dataset. This ia an optional parameter. If you do not specify
a value for output then the added column will take the name of the
specified index with a "z" appended.}

\item{digits}{The number of decimal places for \code{output}. Defaults to 2 d.p.}
}
\value{
A data.frame of the survey dataset with the calculated z-scores added.
}
\description{
Function to add the WHO Growth Reference z-scores to a data frame
}
\examples{
#

}
