% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro1}
\alias{anthro1}
\title{Anthropometric data from a SMART survey in Kabul, Afghanistan.}
\format{
A data frame with 873 observations and 11 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Description} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{age} \tab Age of child (months) \cr
   \emph{sex} \tab Gender of child \cr
   \emph{weight} \tab Weight of child (kgs) \cr
   \emph{height} \tab Height of child (cm) \cr
   \emph{muac} \tab Mid-upper arm circumference (mm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
   \emph{haz} \tab Height-for-age z-score \cr
   \emph{waz} \tab Weight-for-age z-score \cr
   \emph{whz} \tab Weight-for-height z-score \cr
   \emph{flag} \tab Data quality flag \cr
}
}
\usage{
anthro1
}
\description{
Anthropometric data from a SMART survey in Kabul, Afghanistan.
}
\keyword{datasets}
