% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anthro2}
\alias{anthro2}
\title{Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.}
\format{
A data frame with 796 observations and 6 variables\tabular{ll}{
   \strong{Variable} \tab \strong{Definition} \cr
   \emph{psu} \tab Primary sampling unit \cr
   \emph{age} \tab Age (months) \cr
   \emph{sex} \tab Gender \cr
   \emph{wt} \tab Weight (kg) \cr
   \emph{ht} \tab Height (cm) \cr
   \emph{oedema} \tab Presence or absence of oedema \cr
}
}
\usage{
anthro2
}
\description{
Anthropometric data from a single state from a Demographic and Health Survey
(DHS) of a West African country.
}
\keyword{datasets}
