% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\docType{data}
\name{anthro3}
\alias{anthro3}
\title{anthro3}
\format{A data frame with 221 observations and 7 variables
\describe{
\item{\code{psu}}{Primary sampling unit}
\item{\code{age}}{Age (months)}
\item{\code{sex}}{Gender}
\item{\code{weight}}{Weight (kg)}
\item{\code{height}}{Height (cm)}
\item{\code{muac}}{Mid-upper arm circumference (cm)}
\item{\code{oedema}}{Presence or absence of oedema}
}}
\usage{
anthro3
}
\description{
Anthropometric data from a Rapid Assessment Method (RAM) survey from Burundi.
}
\keyword{datasets}
