% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscorer.R
\docType{data}
\name{anthro4}
\alias{anthro4}
\title{A subset of mid-upper arm circumference data from study conducted to create
MUAC-for-age z-scores}
\format{A data.frame with 257 observations and 4 variables
\describe{
  \item{\code{pk_serial}}{Unique identifier}
  \item{\code{muac}}{Mid-upper arm circumference in centimetres}
  \item{\code{agemons}}{Age in months}
  \item{\code{sex}}{Sex; 1 = Male; 2 = Female}
}}
\source{
Mramba Lazarus, Ngari Moses, Mwangome Martha, Muchai Lilian, Bauni
  Evasius, Walker A Sarah et al. A growth reference for mid upper arm
  circumference for age among school age children and adolescents, and
  validation for mortality: growth curve construction and longitudinal
  cohort study BMJ 2017; 358 :j3423
}
\usage{
anthro4
}
\description{
A subset of mid-upper arm circumference data from study conducted to create
MUAC-for-age z-scores
}
\keyword{datasets}
