% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWGS.R
\name{getWGS}
\alias{getWGS}
\title{getWGS}
\usage{
getWGS(sexObserved, firstPart, secondPart, index)
}
\arguments{
\item{sexObserved}{Sex of child (1 = Male; 2 = Female)}

\item{firstPart}{Weight (in kg for WHZ and WAZ) or height (in cm for HAZ) or
head circumference (in cm for )}

\item{secondPart}{Age (months; for HAZ and WAZ) or height (cm; for WHZ)}

\item{index}{One of "wfh", "hfa", "wfa" (specifies the required index)}
}
\value{
z-score of the anthropometric index selected
}
\description{
Calculate z-scores for WHZ, HAZ, WAZ using the WHO Growth Reference (2006)
for a single child data.
}
\examples{
  # apply \\code{getWGS()} to a make believe 52 month old male child with weight of
  # 14.6 kg and height of 98.0 cm
  waz <- getWGS(sexObserved = 1,     # 1 = Male / 2 = Female
                firstPart = 14.6,    # Weight in kilograms
                secondPart = 52,     # Age in whole months
                index = "wfa")       # Anthropometric index (weight-for-age)
  waz

  haz <- getWGS(sexObserved = 1,
                firstPart = 98,      # Height in centimetres
                secondPart = 52,
                index = "hfa")       # Anthropometric index (height-for-age)
  haz

  whz <- getWGS(sexObserved = 1,
                firstPart = 14.6,
                secondPart = 98,
                index = "wfh")       # Anthropometric index (weight-for-height)
  whz

}
